+++ apps/smart_home_server_with_mock_swupdate.cpp
@@ -374,7 +374,8 @@ main(void)
       pthread_cond_wait(&cv, &mutex);
     } else {
       ts.tv_sec = (next_event / OC_CLOCK_SECOND);
-      ts.tv_nsec = (next_event % OC_CLOCK_SECOND) * 1.e09 / OC_CLOCK_SECOND;
+      ts.tv_nsec = static_cast<long>((next_event % OC_CLOCK_SECOND) * 1.e09 /
+                                     OC_CLOCK_SECOND);
       pthread_cond_timedwait(&cv, &mutex, &ts);
     }
     pthread_mutex_unlock(&mutex);
